tests/wpt/tests/html/canvas/element/line-styles/2d.line.join.invalid.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.line.join.invalid</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.line.join.invalid</h1>
<p class="desc">Setting lineJoin to invalid values is ignored</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Setting lineJoin to invalid values is ignored");
_addTest(function(canvas, ctx) {

  ctx.lineJoin = 'bevel'
  _assertSame(ctx.lineJoin, 'bevel', "ctx.lineJoin", "'bevel'");

  ctx.lineJoin = 'bevel';
  ctx.lineJoin = 'invalid';
  _assertSame(ctx.lineJoin, 'bevel', "ctx.lineJoin", "'bevel'");

  ctx.lineJoin = 'bevel';
  ctx.lineJoin = 'ROUND';
  _assertSame(ctx.lineJoin, 'bevel', "ctx.lineJoin", "'bevel'");

  ctx.lineJoin = 'bevel';
  ctx.lineJoin = 'round\0';
  _assertSame(ctx.lineJoin, 'bevel', "ctx.lineJoin", "'bevel'");

  ctx.lineJoin = 'bevel';
  ctx.lineJoin = 'round ';
  _assertSame(ctx.lineJoin, 'bevel', "ctx.lineJoin", "'bevel'");

  ctx.lineJoin = 'bevel';
  ctx.lineJoin = "";
  _assertSame(ctx.lineJoin, 'bevel', "ctx.lineJoin", "'bevel'");

  ctx.lineJoin = 'bevel';
  ctx.lineJoin = 'butt';
  _assertSame(ctx.lineJoin, 'bevel', "ctx.lineJoin", "'bevel'");

});
</script>
